backend/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
psycopg2-binary==2.9.5
pillow==10.0.0
drf-yasg==1.21.7    
django-filter==23.2
django-cors-headers==4.2.0
Faker==19.6.2
